♻️ refactor : 코드분리
Post추가 검증 로직 lib/post.ts로 이동
